+++ src/components/Holidays.js
@@ -1,0 +1,281 @@
+export const Holidays = [
+	{
+		_id: 33,
+		Date: "2018-01-01",
+		"Holiday Name": "New Year's Day",
+		Information:
+			"New Year's Day is the first day of the calendar year and is celebrated each January 1st",
+		"More Information": "",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 34,
+		Date: "2018-01-26",
+		"Holiday Name": "Australia Day",
+		Information: "Always celebrated on 26 January",
+		"More Information": "",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 35,
+		Date: "2018-02-12",
+		"Holiday Name": "Royal Hobart Regatta",
+		Information: "",
+		"More Information": "",
+		"Applicable To": "TAS",
+	},
+	{
+		_id: 36,
+		Date: "2018-03-05",
+		"Holiday Name": "Labour Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information":
+			"http:--www.commerce.wa.gov.au-labour-relations-public-holidays-western-australia",
+		"Applicable To": "WA",
+	},
+	{
+		_id: 37,
+		Date: "2018-03-12",
+		"Holiday Name": "Eight Hours Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information": "",
+		"Applicable To": "TAS",
+	},
+	{
+		_id: 38,
+		Date: "2018-03-12",
+		"Holiday Name": "Canberra Day",
+		Information:
+			"Held on the second Monday of March each year in Canberra, to celebrate the naming ceremony of Australia's capital which took place on 12 March 1913.",
+		"More Information": "http:--www.cmd.act.gov.au-communication-holidays",
+		"Applicable To": "ACT",
+	},
+	{
+		_id: 39,
+		Date: "2018-03-12",
+		"Holiday Name": "March public holiday",
+		Information:
+			"The Holidays Act 1910 provides for the third Monday in March to be a public holiday.",
+		"More Information":
+			"http:--www.safework.sa.gov.au-show_page.jsp?id=2483#.U6tbuSInrho",
+		"Applicable To": "SA",
+	},
+	{
+		_id: 40,
+		Date: "2018-03-12",
+		"Holiday Name": "Labour Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information":
+			"http:--www.business.vic.gov.au-victorian-public-holidays-and-daylight-saving-victorian-public-holidays",
+		"Applicable To": "VIC",
+	},
+	{
+		_id: 41,
+		Date: "2018-03-30",
+		"Holiday Name": "Good Friday",
+		Information:
+			"Easter is celebrated with Good Friday and Easter Monday creating a 4 day long weekend.",
+		"More Information": "",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 42,
+		Date: "2018-03-31",
+		"Holiday Name": "Easter Saturday",
+		Information:
+			"Easter Saturday is between Good Friday and Easter Sunday in Australia.",
+		"More Information": "",
+		"Applicable To": "ACT|NSW|NT|QLD|SA|VIC",
+	},
+	{
+		_id: 43,
+		Date: "2018-04-01",
+		"Holiday Name": "Easter Sunday",
+		Information: "Public Holiday as part of Easter.",
+		"More Information": "",
+		"Applicable To": "ACT|NSW|QLD|VIC",
+	},
+	{
+		_id: 44,
+		Date: "2018-04-02",
+		"Holiday Name": "Easter Monday",
+		Information: "Public Holiday as part of Easter.",
+		"More Information": "",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 45,
+		Date: "2018-04-25",
+		"Holiday Name": "ANZAC Day",
+		Information: "Celebrated on the 25 April each year.",
+		"More Information": "http:--www.awm.gov.au-commemoration-anzac-day-",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 46,
+		Date: "2018-05-07",
+		"Holiday Name": "May Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information":
+			"http:--www.ocpe.nt.gov.au-working_in_the_ntps-legislation-public_holidays%0A",
+		"Applicable To": "NT",
+	},
+	{
+		_id: 47,
+		Date: "2018-05-07",
+		"Holiday Name": "Labour Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information": "",
+		"Applicable To": "QLD",
+	},
+	{
+		_id: 48,
+		Date: "2018-05-28",
+		"Holiday Name": "Reconciliation Day",
+		Information: "",
+		"More Information": "",
+		"Applicable To": "ACT",
+	},
+	{
+		_id: 49,
+		Date: "2018-06-04",
+		"Holiday Name": "Western Australia Day",
+		Information:
+			"Held on the first Monday in June each year and is a state holiday only.",
+		"More Information":
+			"http:--www.commerce.wa.gov.au-labour-relations-public-holidays-western-australia",
+		"Applicable To": "WA",
+	},
+	{
+		_id: 50,
+		Date: "2018-06-11",
+		"Holiday Name": "Queen's Birthday",
+		Information:
+			"Celebrated on second Monday in June except in Western Australia and Queensland.",
+		"More Information":
+			"https:--www.gg.gov.au-queens-birthday-2014-honours-list",
+		"Applicable To": "ACT|NSW|NT|SA|TAS|VIC",
+	},
+	{
+		_id: 51,
+		Date: "2018-08-06",
+		"Holiday Name": "Picnic Day",
+		Information:
+			"Picnic Day is observed in Australia’s Northern Territory on the first Monday of August each year. It serves as an extra day of rest for workers as well as part of a long weekend for various events, such as the Harts Range annual races.",
+		"More Information":
+			"http:--www.ocpe.nt.gov.au-working_in_the_ntps-legislation-public_holidays%0A",
+		"Applicable To": "NT",
+	},
+	{
+		_id: 52,
+		Date: "2018-08-06",
+		"Holiday Name": "Bank Holiday",
+		Information: "",
+		"More Information": "",
+		"Applicable To": "NSW",
+	},
+	{
+		_id: 53,
+		Date: "2018-09-24",
+		"Holiday Name": "Queen's Birthday",
+		Information:
+			"Non-metropolitan local authorities in WA may celebrate the Queen's Birthday Public Holiday on an alternative date",
+		"More Information":
+			"http:--www.commerce.wa.gov.au-labourrelations-content-Employers-Pay_rates_leave_and_conditions-Regional_Dates_for_the_Queens.html",
+		"Applicable To": "WA",
+	},
+	{
+		_id: 54,
+		Date: "2018-09-28",
+		"Holiday Name": "Friday before the AFL Grand Final",
+		Information: "Friday before the AFL Grand Final",
+		"More Information": "",
+		"Applicable To": "VIC",
+	},
+	{
+		_id: 55,
+		Date: "2018-10-01",
+		"Holiday Name": "Labour Day",
+		Information:
+			"Always on a Monday, creating a long weekend. It celebrates the eight-hour working day, a victory for workers in the mid-late 19th century.",
+		"More Information": "",
+		"Applicable To": "ACT|NSW|SA",
+	},
+	{
+		_id: 56,
+		Date: "2018-10-01",
+		"Holiday Name": "Queen's Birthday",
+		Information: "Celebrated on First Monday in October in Queensland.",
+		"More Information":
+			"https:--www.gg.gov.au-queens-birthday-2014-honours-list",
+		"Applicable To": "QLD",
+	},
+	{
+		_id: 57,
+		Date: "2018-11-05",
+		"Holiday Name": "Recreation Day",
+		Information: "",
+		"More Information": "",
+		"Applicable To": "TAS",
+	},
+	{
+		_id: 58,
+		Date: "2018-11-06",
+		"Holiday Name": "Melbourne Cup",
+		Information:
+			"All of Victoria unless alternate local holiday has been arranged by non-metro council.",
+		"More Information":
+			"http:--www.business.vic.gov.au-victorian-public-holidays-and-daylight-saving-victorian-public-holidays",
+		"Applicable To": "VIC",
+	},
+	{
+		_id: 59,
+		Date: "2018-12-24",
+		"Holiday Name": "Christmas Eve",
+		Information: "Part-day public holiday from 7pm-12 midnight",
+		"More Information":
+			"http:--www.safework.sa.gov.au-show_page.jsp?id=2483",
+		"Applicable To": "NT|SA",
+	},
+	{
+		_id: 60,
+		Date: "2018-12-25",
+		"Holiday Name": "Christmas Day",
+		Information:
+			"Christmas Day is an annual holiday which celebrates the birth of Jesus Christ over 2000 years ago.",
+		"More Information": "",
+		"Applicable To": "NATIONAL",
+	},
+	{
+		_id: 61,
+		Date: "2018-12-26",
+		"Holiday Name": "Boxing Day",
+		Information:
+			"Boxing Day occurs the day after Christmas. Sydney-to-Hobart yacht race and Boxing Day Test Match (Cricket) start on this day.",
+		"More Information": "",
+		"Applicable To": "ACT|NSW|NT|QLD|TAS|VIC|WA",
+	},
+	{
+		_id: 62,
+		Date: "2018-12-26",
+		"Holiday Name": "Proclamation Day",
+		Information: "Occurs the day after Christmas Day.",
+		"More Information":
+			"http:--www.safework.sa.gov.au-show_page.jsp?id=2483#.U6tbuSInrho",
+		"Applicable To": "SA",
+	},
+	{
+		_id: 63,
+		Date: "2018-12-31",
+		"Holiday Name": "New Year's Eve",
+		Information: "Part-day public holiday from 7pm-12 midnight",
+		"More Information":
+			"http:--www.safework.sa.gov.au-show_page.jsp?id=2483",
+		"Applicable To": "NT|SA",
+	},
+];