+++ src/pages/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 
+import Campaign from "../components/Campaign";
 import Card from "../components/Card";
 import Campaign from "../components/Campaign";